requirements.txt
# Python 3.7+ ile uyumlu sürümler
PyQt5>=5.15.0,<5.16.0
pyqtgraph>=0.12.0
pyserial>=3.5
numpy>=1.19.0; python_version<"3.8"
numpy>=1.20.0; python_version>="3.8" and python_version<"3.9"
numpy>=1.21.0; python_version>="3.9" and python_version<"3.11"
numpy>=1.23.0; python_version>="3.11" and python_version<"3.12"
numpy>=1.26.0; python_version>="3.12"
pandas>=1.3.0; python_version<"3.8"
pandas>=1.4.0; python_version>="3.8" and python_version<"3.9"
pandas>=1.5.0; python_version>="3.9" and python_version<"3.11"
pandas>=1.5.0; python_version>="3.11" and python_version<"3.12"
pandas>=2.0.0; python_version>="3.12"
matplotlib>=3.3.0; python_version<"3.8"
matplotlib>=3.5.0; python_version>="3.8" and python_version<"3.11"
matplotlib>=3.6.0; python_version>="3.11" and python_version<"3.12"
matplotlib>=3.7.0; python_version>="3.12"